Added Hud Panels Data and Polish

+++ AoE_Project/Motor2D/Hud_GamePanel.cpp
@@ -331,7 +331,7 @@ bool Game_Panel::AddResource(int amount, PLAYER_RESOURCES resource_type)
 
 bool Game_Panel::UseResource(int amount_wood, int amount_food, int amount_gold, int amount_stone, int used_population)
 {
-	if(wood-amount_wood >= 0 && meat-amount_food >= 0 && gold - amount_gold >= 0 && stone - amount_stone >= 0 &&  population + used_population <= max_population)
+	if(wood-amount_wood >= 0 && meat-amount_food >= 0 && gold - amount_gold >= 0 && stone - amount_stone >= 0 &&  population + used_population <= (int)max_population)
 	{
 		if (amount_wood != 0) AddResource(-amount_wood, GP_WOOD);
 		if (amount_food != 0) AddResource(-amount_food, GP_MEAT);
